requirements.txt
mutagen>=1.47.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
requests>=2.31.0
croniter>=2.0.1
Pillow>=10.0.0

